Move tests for users from groups to users integration tests MODUSERS-309
