+++ src/test/java/org/folio/moduserstest/GroupIT.java
@@ -356,61 +356,4 @@ class GroupIT {
 
     assertThat(usersFilteredByGroupName.getTotalRecords(), is(0));
   }
-
-  //These tests should  be in the integration tests for users not groups
-  //they can be moved when the users integration tests are improved
-  @Test
-  void canFindActiveUsers() {
-    usersClient.createUser(User.builder()
-      .username("steve")
-      .active(true)
-      .build());
-
-    usersClient.createUser(User.builder()
-      .username("joanne")
-      .active(false)
-      .build());
-
-    usersClient.createUser(User.builder()
-      .username("jenna")
-      .active(true)
-      .build());
-
-    final var activeUsers = usersClient.getUsers("active=true");
-
-    assertThat(activeUsers.getTotalRecords(), is(2));
-  }
-
-  @Test
-  void cannotCreateUserWithSameUsernameAsExistingUser() {
-    usersClient.createUser(User.builder()
-      .username("julia").build());
-
-    final var response = usersClient.attemptToCreateUser(User.builder()
-      .username("julia").build());
-
-    response.statusCode(is(422));
-
-    final var errors = response.extract().as(ValidationErrors.class);
-
-    assertThat(errors.getErrors().get(0).getMessage(),
-      is("User with this username already exists"));
-  }
-
-  @Test
-  void cannotCreateUserWithSameIdAsExistingUser() {
-    final var existingUser = usersClient.createUser(User.builder()
-      .username("julia").build());
-
-    final var response = usersClient.attemptToCreateUser(User.builder()
-      .id(existingUser.getId())
-      .username("steve").build());
-
-    response.statusCode(is(422));
-
-    final var errors = response.extract().as(ValidationErrors.class);
-
-    assertThat(errors.getErrors().get(0).getMessage(),
-      is("User with this id already exists"));
-  }
 }
